+++ packages/esm-amd-loader/CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-# Unreleased
+<!-- # Unreleased -->
+<!-- Add new, unreleased changes here. -->
+
+## [1.0.0] - 2018-05-08
 * Fix bug where the ordering of module execution was not strict enough.
   Standard ES modules are loaded as they're encountered, but they execute
   strictly in the order that they're first imported.
 * Support cyclical dependencies.
-<!-- Add new, unreleased changes here. -->
 
 ## [0.1.1] - 2018-05-03
 * Fix bug where the `import.meta.url` of a top-level module had an artificial